+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 
 * Fixed map issue under Py3
 * Replaced numpy.random with a RandomState
+* Restart sampling feature
 
 0.1.2 (2016-01-27)
 ++++++++++++++++++